requirements.txt
tensorboard==2.10.1
tokenizers==0.12.1
torchmetrics==0.9.0
deepspeed==0.7.3
transformers==4.23.1
pandas==1.5.0